chore: dedupe levenshteinDistance into a shared util
Four copies of the Levenshtein distance function existed across the
codebase (schema.ts, validation.ts, template.ts, and an exported one in
discovery.ts). Extract a single implementation into src/lib/levenshtein.ts
and update all call sites to import from it. The implementations were all
mathematically equivalent (template.ts swapped the a/b matrix axes, which
is harmless since Levenshtein distance is symmetric).

Move the unit tests off discovery.test.ts into a dedicated
levenshtein.test.ts and expand coverage.

Closes #604
